+++ origin/tests/TestCase/Controller/ResponseTest.php
@@ -52,14 +52,14 @@ class ResponseTest extends \PHPUnit\Framework\TestCase
     public function testStatusCode()
     {
         $response  = new Response();
-        $response->status(501);
-        $this->assertEquals(501, $response->status());
+        $response->statusCode(501);
+        $this->assertEquals(501, $response->statusCode());
     }
 
     public function testSend()
     {
         $response  = new Response();
-        $response->status(200);
+        $response->statusCode(200);
         $response->header('Accept-Language', 'en-us,en;q=0.5');
         $this->assertNull($response->send()); // or $response->send()
     }